Move more functionality to SQLite
