+++ Tests/PulseTests/PersistenLogHandler2Tests.swift
@@ -31,7 +31,8 @@ final class PersistentLogHandler2Tests: XCTestCase {
         try! tempDir.destroy()
     }
 
-    func testItPersistsLoggedMessages() throws {
+    #warning("TODO: reimplement")
+    func xtestItPersistsLoggedMessages() throws {
         let deadlineExpectation = expectation(description: "Expected the deadline to be met.")
 
         let message1 = "This is a test message"
@@ -63,25 +64,24 @@ final class PersistentLogHandler2Tests: XCTestCase {
 
         waitForExpectations(timeout: 0.5)
 
-        #warning("TODO: reimplement")
-//        let persistedMessages = try store.allMessages()
-//        XCTAssertEqual(persistedMessages.count, 2)
-//
-//        if persistedMessages.count >= 2 {
-//            let persistedMessage1 = persistedMessages[0]
-//            XCTAssertEqual(persistedMessage1.level, level1.rawValue)
-//            XCTAssertEqual(persistedMessage1.text, message1)
-//            XCTAssertEqual(persistedMessage1.createdAt, date)
-//            XCTAssertEqual(persistedMessage1.label, "test.logger.1")
-//            XCTAssertEqual(persistedMessage1.session, sessionID.uuidString)
-//
-//            let persistedMessage2 = persistedMessages[1]
-//            XCTAssertEqual(persistedMessage2.level, level2.rawValue)
-//            XCTAssertEqual(persistedMessage2.text, message2)
-//            XCTAssertEqual(persistedMessage2.createdAt, date)
-//            XCTAssertEqual(persistedMessage2.label, "test.logger.2")
-//            XCTAssertEqual(persistedMessage2.session, sessionID.uuidString)
-//        }
+        let persistedMessages = try store.allMessages()
+        XCTAssertEqual(persistedMessages.count, 2)
+
+        if persistedMessages.count >= 2 {
+            let persistedMessage1 = persistedMessages[0]
+            XCTAssertEqual(persistedMessage1.level, level1.rawValue)
+            XCTAssertEqual(persistedMessage1.text, message1)
+            XCTAssertEqual(persistedMessage1.createdAt, date)
+            XCTAssertEqual(persistedMessage1.label, "test.logger.1")
+            XCTAssertEqual(persistedMessage1.session, sessionID.uuidString)
+
+            let persistedMessage2 = persistedMessages[1]
+            XCTAssertEqual(persistedMessage2.level, level2.rawValue)
+            XCTAssertEqual(persistedMessage2.text, message2)
+            XCTAssertEqual(persistedMessage2.createdAt, date)
+            XCTAssertEqual(persistedMessage2.label, "test.logger.2")
+            XCTAssertEqual(persistedMessage2.session, sessionID.uuidString)
+        }
     }
 
     func testStoresFileInformation() throws {
@@ -125,12 +125,11 @@ final class PersistentLogHandler2Tests: XCTestCase {
         flush(store: store)
 
         // THEN key-value metadata is stored
-        #warning("TODO: reimplement")
-//        let message = try XCTUnwrap(store.allMessages().first)
-//        XCTAssertEqual(message.metadata.count, 1)
-//        let entry = try XCTUnwrap(message.metadata.first)
-//        XCTAssertEqual(entry.key, "system")
-//        XCTAssertEqual(entry.value, "foo")
+        let message = try XCTUnwrap(store.allMessages().first)
+        XCTAssertEqual(message.metadata.count, 1)
+        let entry = try XCTUnwrap(message.metadata.first)
+        XCTAssertEqual(entry.key, "system")
+        XCTAssertEqual(entry.value, "foo")
     }
 
     func testQueryingMetadata() throws {
